NLP Course/Class 26/requirements.txt
streamlit
requests
llama-index
matplotlib
scikit-image
pillow
langchain-community